Reorganize all subprotocol dirs
