+++ packages/portalnetwork/src/subprotocols/history/util.ts
@@ -1,7 +1,7 @@
 import SHA256 from '@chainsafe/as-sha256'
 import { fromHexString, toHexString } from '@chainsafe/ssz'
 import { HistoryNetworkContentKeyUnionType } from '.'
-import { PortalNetwork } from '..'
+import { PortalNetwork } from '../..'
 import { HistoryNetworkContentTypes } from './types'
 import * as rlp from 'rlp'
 import { Block, BlockBuffer } from '@ethereumjs/block'
