+++ packages/portalnetwork/src/subprotocols/state/util.ts
